IDBOOKAPI/requirements.txt
amqp
arabic-reshaper
asgiref
asn1crypto
async-timeout
attrs
billiard
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
coreapi
coreschema
cryptography
cssselect2
djangify
django
django-appconf
django-cors-headers
django-environ
# django-faker
django-filter
django-imagekit
django-localflavor
django-pytest
djangorestframework
djangorestframework-simplejwt
drf-tracking
drf-yasg
exceptiongroup
fake-factory
Faker
h11
h2
hpack
hstspreload
html5lib
httpx
hyperframe
idna
imagekitio
inflection
iniconfig
itypes
Jinja2
kombu
lxml
MarkupSafe
model-bakery
msg91-otp
oscrypto
outcome
packaging
pdfkit
pilkit
Pillow
pluggy
prompt-toolkit
psycopg2-binary
pycparser
pyfcm
pyHanko
pyhanko-certvalidator
pyhtml2pdf
PyJWT
pypdf
pypng
PySocks
pytest
pytest-csv
python-bidi
python-dateutil
python-dotenv
python-stdnum
pytz
PyYAML
qrcode
razorpay
redis
reportlab
requests
requests-toolbelt
rfc3986
selenium
six
sniffio
sortedcontainers
sqlparse
svglib
tinycss2
tomli
tqdm
trio
trio-websocket
typing_extensions
tzdata
tzlocal
uritemplate
uritools
urllib3
vine
wcwidth
webdriver-manager
webencodings
wsproto
xhtml2pdf
gunicorn
